// src/features/movies/types.ts
export interface MovieStructure {
  id: number;
  title: string;
  movie_banner: string;
  description: string;
  director: string;
  release_date: string;
  running_time: string;
}

export interface MovieStructureFiltered {
  id: number;
  title: string;
  movieBanner: string;
  description: string;
  director: string;
  releaseDate: string;
  runningTime: string;
}